+++ spec/pump/xml/value_spec.rb
@@ -15,7 +15,7 @@ describe Pump::Xml::Value do
   end
 
   describe "#to_s" do
-    its(:to_s) { should eql("\#{object.method_name.to_s.encode(:xml => :text)}") }
+    its(:to_s) { should eql("\#{remove_ilegal_chars object.method_name.to_s.encode(:xml => :text)}") }
 
     context "with option :xmlsafe => true" do
       subject { Pump::Xml::Value.new("method_name", {}, [], :xmlsafe => true) }
@@ -23,9 +23,15 @@ describe Pump::Xml::Value do
       its(:to_s) { should eql("\#{object.method_name}") }
     end
 
+    context "with option :typecast => :xmlschema" do
+      subject { Pump::Xml::Value.new("method_name", {}, [], :typecast => :xmlschema) }
+
+      its(:to_s) { should eql("\#{object.method_name.xmlschema}") }
+    end
+
     context "with path name" do
       it do
-        subject.to_s('custom_path').should eql("\#{custom_path.to_s.encode(:xml => :text)}")
+        subject.to_s('custom_path').should eql("\#{remove_ilegal_chars custom_path.to_s.encode(:xml => :text)}")
       end
     end
   end